+++ R/rand_forest-data.R
@@ -224,4 +224,86 @@ make_rand_forest_aorsf <- function() {
   )
 }
 
+make_rand_forest_ranger <- function() {
+  parsnip::set_model_engine(
+    "rand_forest",
+    mode = "censored regression",
+    eng = "ranger"
+  )
+  parsnip::set_dependency(
+    "rand_forest",
+    eng = "ranger",
+    pkg = "ranger",
+    mode = "censored regression"
+  )
+  parsnip::set_dependency(
+    "rand_forest",
+    eng = "ranger",
+    pkg = "censored",
+    mode = "censored regression"
+  )
+
+  parsnip::set_fit(
+    model = "rand_forest",
+    eng = "ranger",
+    mode = "censored regression",
+    value = list(
+      interface = "formula",
+      data = c(formula = "formula", data = "data", weights = "case.weights"),
+      protect = c("formula", "data", "weights"),
+      func = c(pkg = "ranger", fun = "ranger"),
+      defaults = list(
+        num.threads = 1,
+        verbose = FALSE,
+        seed = rlang::expr(sample.int(10^5, 1))
+      )
+    )
+  )
+
+  parsnip::set_encoding(
+    model = "rand_forest",
+    mode = "censored regression",
+    eng = "ranger",
+    options = list(
+      predictor_indicators = "none",
+      compute_intercept = FALSE,
+      remove_intercept = FALSE,
+      allow_sparse_x = FALSE
+    )
+  )
+
+  parsnip::set_pred(
+    model = "rand_forest",
+    eng = "ranger",
+    mode = "censored regression",
+    type = "time",
+    value = list(
+      pre = NULL,
+      post = NULL,
+      func = c(pkg = "censored", fun = "survival_time_ranger"),
+      args = list(
+        object = rlang::expr(object),
+        new_data = rlang::expr(new_data)
+      )
+    )
+  )
+
+  parsnip::set_pred(
+    model = "rand_forest",
+    eng = "ranger",
+    mode = "censored regression",
+    type = "survival",
+    value = list(
+      pre = NULL,
+      post = NULL,
+      func = c(pkg = "censored", fun = "survival_prob_ranger"),
+      args = list(
+        object = rlang::expr(object),
+        new_data = rlang::expr(new_data),
+        eval_time = rlang::expr(eval_time)
+      )
+    )
+  )
+}
+
 # nocov end